feat: 诊间支付参数封装 9.14

See merge request server1/hsa-his!2753

+++ hsa-com/src/main/java/cn/hsa/module/payment/entity/PaymentOrderDO.java
@@ -1,5 +1,6 @@
 package cn.hsa.module.payment.entity;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -56,11 +57,11 @@ public class PaymentOrderDO implements Serializable {
     /**
      * 总费用
      */
-    private Double totalPrice;
+    private BigDecimal totalPrice;
     /**
      * 平台支付费用
      */
-    private Double paymentPrice;
+    private BigDecimal paymentPrice;
     /**
      * 支付方式（ZFFS）
      */
